Create javalogic.html
Logic for the java part of the website. Will be getting inputs from prompts rather than from the website, and it will need to be changed once logic is added. Grades will be only added in strings without spaces (ex. A+ will run but A + will not), and retakes will be specified with a "Y" or "N"